Move PostProcessing to Video
